+++ src/components/RTable/src/components/Fullscreen.tsx
@@ -13,7 +13,7 @@ import { NPopover } from 'naive-ui'
 import RIcon from '@/components/RIcon/index'
 
 import config from '../config'
-import { useFullscreen } from '@/hooks/web/index'
+import { useFullscreen } from 'vue-hooks-plus'
 
 import type { TableProvider } from '../type'
 
@@ -24,14 +24,18 @@ export default defineComponent({
       config.tableKey,
       {} as TableProvider,
     )
-    const [isFullscreen, { toggleFullscreen }] = useFullscreen(wrapperRef)
+    const [isFullscreen, { toggleFullscreen, isEnabled }] =
+      useFullscreen(wrapperRef)
 
     return {
       toggleFullscreen,
       isFullscreen,
+      isEnabled,
     }
   },
   render() {
+    const { toggleFullscreen, isEnabled, $t } = this
+
     return (
       <NPopover showArrow={false}>
         {{
@@ -40,7 +44,13 @@ export default defineComponent({
               name="fullscreen"
               size={config.tableIconSize}
               cursor="pointer"
-              onClick={this.toggleFullscreen.bind(this)}
+              onClick={() => {
+                if (!isEnabled) {
+                  $t('globalMessage.isEnabledFullscreen')
+                }
+
+                toggleFullscreen()
+              }}
             />
           ),
           default: () => (this.isFullscreen ? '取消全屏' : '全屏表格'),