+++ src/utils/vue/index.ts
@@ -1,2 +1,3 @@
 export { call } from './call'
 export { unrefElement } from './unrefElement'
+export { renderNode } from './renderNode'